fix some alert

+++ Zju_SmartHomeV1/Zju_SmartHome/Classes/DLLampControlDinnerYWModeViewController.m
@@ -107,7 +107,7 @@
   [slider addTarget:self action:@selector(sliderValueChanged) forControlEvents:UIControlEventValueChanged];
   
   
-  if (fabsf(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
+  if (fabs(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
     // 5 & 5s & 5c
     imgView.image = [UIImage imageNamed:@"YWCircle_5"];
     viewColorPickerPositionIndicator.frame = CGRectMake(70, 70, 16, 16);
@@ -116,7 +116,7 @@
     btnPlay.frame = CGRectMake(111, 111, 60, 60);
     slider.frame = CGRectMake(40, 260, 200, 10);
     
-  }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
+  }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
     // 6 & 6s
     imgView.image = [UIImage imageNamed:@"YWCircle_6"];
     viewColorPickerPositionIndicator.frame = CGRectMake(75, 75, 20, 20);
@@ -125,7 +125,7 @@
     btnPlay.frame = CGRectMake(135, 135, 60, 60);
     slider.frame = CGRectMake(50, 310, 225, 10);
     
-  }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
+  }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
     // 6p & 6sp
     imgView.image = [UIImage imageNamed:@"YWCircle_6p"];
     viewColorPickerPositionIndicator.frame = CGRectMake(80, 80, 24, 24);
@@ -251,7 +251,7 @@
        *  从零取到一百，生生的把它凑成了一百
        */
       int cwValue = (int)(touchLocation.y / 2.5) - 2;
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
         // 4 & 4s
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -259,7 +259,7 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
       }
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
         // 5 & 5s & 5c
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -267,7 +267,7 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
         
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
         // 6 & 6s
         if (cwValue < 63) {
           cwValue = cwValue;
@@ -275,7 +275,7 @@
           cwValue += 1;
         }
         
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
         // 6p & 6sp
         
         if (cwValue < 63) {
@@ -339,7 +339,7 @@
       
       
       int cwValue = (int)(touchLocation.y / 2.5) - 2;
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
         // 4 & 4s
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -347,21 +347,21 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
       }
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
         // 5 & 5s & 5c
         if (cwValue < 63) {
           cwValue = cwValue + 1;
         }else{
           cwValue = (float)(cwValue) / 81 * 100;
         }
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
         // 6 & 6s
         if (cwValue < 63) {
           cwValue = cwValue;
         }else{
           cwValue += 1;
         }
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
         // 6p & 6sp
         if (cwValue < 63) {
           cwValue = cwValue;
@@ -433,7 +433,7 @@
        *  从零取到一百，生生的把它凑成了一百
        */
       int cwValue = (int)(touchLocation.y / 2.5) - 2;
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
         // 4 & 4s
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -441,7 +441,7 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
       }
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
         // 5 & 5s & 5c
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -449,7 +449,7 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
         
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
         // 6 & 6s
         if (cwValue < 63) {
           cwValue = cwValue;
@@ -457,7 +457,7 @@
           cwValue += 1;
         }
         
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
         // 6p & 6sp
         
         if (cwValue < 63) {
