+++ Zju_SmartHomeV1/Zju_SmartHome/Classes/DLLampControllYWModeViewController.m
@@ -106,7 +106,7 @@
   [slider addTarget:self action:@selector(sliderValueChanged) forControlEvents:UIControlEventValueChanged];
   
   
-  if (fabsf(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
+  if (fabs(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
     // 5 & 5s & 5c
     imgView.image = [UIImage imageNamed:@"YWCircle_5"];
     viewColorPickerPositionIndicator.frame = CGRectMake(70, 70, 16, 16);
@@ -115,7 +115,7 @@
     btnPlay.frame = CGRectMake(111, 111, 60, 60);
     slider.frame = CGRectMake(40, 260, 200, 10);
     
-  }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
+  }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
     // 6 & 6s
     imgView.image = [UIImage imageNamed:@"YWCircle_6"];
     viewColorPickerPositionIndicator.frame = CGRectMake(75, 75, 20, 20);
@@ -124,7 +124,7 @@
     btnPlay.frame = CGRectMake(135, 135, 60, 60);
     slider.frame = CGRectMake(50, 310, 225, 10);
     
-  }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
+  }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
     // 6p & 6sp
     imgView.image = [UIImage imageNamed:@"YWCircle_6p"];
     viewColorPickerPositionIndicator.frame = CGRectMake(80, 80, 24, 24);
@@ -245,7 +245,7 @@
        *  从零取到一百，生生的把它凑成了一百
        */
       int cwValue = (int)(touchLocation.y / 2.5) - 2;
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
         // 4 & 4s
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -253,7 +253,7 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
       }
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
         // 5 & 5s & 5c
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -261,7 +261,7 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
         
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
         // 6 & 6s
         if (cwValue < 63) {
           cwValue = cwValue;
@@ -269,7 +269,7 @@
           cwValue += 1;
         }
         
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
         // 6p & 6sp
         
         if (cwValue < 63) {
@@ -333,7 +333,7 @@
       
       
       int cwValue = (int)(touchLocation.y / 2.5) - 2;
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
         // 4 & 4s
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -341,21 +341,21 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
       }
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
         // 5 & 5s & 5c
         if (cwValue < 63) {
           cwValue = cwValue + 1;
         }else{
           cwValue = (float)(cwValue) / 81 * 100;
         }
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
         // 6 & 6s
         if (cwValue < 63) {
           cwValue = cwValue;
         }else{
           cwValue += 1;
         }
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
         // 6p & 6sp
         if (cwValue < 63) {
           cwValue = cwValue;
@@ -429,7 +429,7 @@
        *  从零取到一百，生生的把它凑成了一百
        */
       int cwValue = (int)(touchLocation.y / 2.5) - 2;
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 480)) < 1) {
         // 4 & 4s
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -437,7 +437,7 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
       }
-      if (fabsf(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
+      if (fabs(([[UIScreen mainScreen] bounds].size.height - 568)) < 1){
         // 5 & 5s & 5c
         if (cwValue < 63) {
           cwValue = cwValue + 1;
@@ -445,7 +445,7 @@
           cwValue = (float)(cwValue) / 81 * 100;
         }
         
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 667)) < 1) {
         // 6 & 6s
         if (cwValue < 63) {
           cwValue = cwValue;
@@ -453,7 +453,7 @@
           cwValue += 1;
         }
         
-      }else if (fabsf(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
+      }else if (fabs(([[UIScreen mainScreen] bounds].size.height - 736)) < 1){
         // 6p & 6sp
         
         if (cwValue < 63) {
